requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
mysql-connector-python==8.0.17
mysqlclient==2.0.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
sqlparse==0.4.1
